Refactor code to fix all SwiftLint errors
Split multi-type files into individual files following one-type-per-file convention:
- Split 7 test files into 24 individual test suites
- Split WikiAPITypes into ParseResponse, ParseContent, TextContent
- Split MockDataSourceFetchers into 6 individual mock files
- Rename FieldValueAssertions to FieldValue+Assertions

Reduce cyclomatic complexity by extracting helper methods:
- MockCloudKitService.executeBatchOperations: Extract 5 operation handlers (complexity 17 → 2)
- DataSourcePipeline.fetchRestoreImages: Extract 6 data source fetchers (complexity 18 → 2)
- DataSourcePipeline.mergeRestoreImages: Extract 4 merge helpers (complexity 17 → 3)

Add SwiftLint exceptions for acceptable cases:
- Test fixture identifiers (preserve readable names like sonoma14_2_1)
- Large test files with comprehensive coverage
- Configuration files with extensive validation logic

All SwiftLint errors fixed. Linting now completes successfully.
