+++ Sources/ConfigKeyKit/ConfigurationKey.swift
@@ -1,11 +1,35 @@
 //
 //  ConfigurationKey.swift
-//  ConfigKeyKit
+//  BushelCloud
 //
-//  Generic configuration key infrastructure with naming style support
+//  Created by Leo Dion.
+//  Copyright © 2025 BrightDigit.
+//
+//  Permission is hereby granted, free of charge, to any person
+//  obtaining a copy of this software and associated documentation
+//  files (the "Software"), to deal in the Software without
+//  restriction, including without limitation the rights to use,
+//  copy, modify, merge, publish, distribute, sublicense, and/or
+//  sell copies of the Software, and to permit persons to whom the
+//  Software is furnished to do so, subject to the following
+//  conditions:
+//
+//  The above copyright notice and this permission notice shall be
+//  included in all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+//  OTHER DEALINGS IN THE SOFTWARE.
 //
 
 import Foundation
+
+// swiftlint:disable file_length
 
 // MARK: - Configuration Key Source
 
@@ -119,7 +143,7 @@ public struct ConfigKey<Value: Sendable>: ConfigurationKey, Sendable {
     self.baseKey = base
     self.styles = [
       .commandLine: StandardNamingStyle.dotSeparated,
-      .environment: StandardNamingStyle.screamingSnakeCase(prefix: envPrefix)
+      .environment: StandardNamingStyle.screamingSnakeCase(prefix: envPrefix),
     ]
     self.explicitKeys = [:]
     self.defaultValue = defaultVal
@@ -209,7 +233,7 @@ public struct OptionalConfigKey<Value: Sendable>: ConfigurationKey, Sendable {
     self.baseKey = base
     self.styles = [
       .commandLine: StandardNamingStyle.dotSeparated,
-      .environment: StandardNamingStyle.screamingSnakeCase(prefix: envPrefix)
+      .environment: StandardNamingStyle.screamingSnakeCase(prefix: envPrefix),
     ]
     self.explicitKeys = [:]
   }
@@ -274,7 +298,7 @@ extension ConfigKey where Value == Bool {
     self.baseKey = base
     self.styles = [
       .commandLine: StandardNamingStyle.dotSeparated,
-      .environment: StandardNamingStyle.screamingSnakeCase(prefix: envPrefix)
+      .environment: StandardNamingStyle.screamingSnakeCase(prefix: envPrefix),
     ]
     self.explicitKeys = [:]
     self.defaultValue = defaultVal
