+++ Tests/BushelCloudKitTests/DataSources/RestoreImageMergeTests.swift
@@ -1,5 +1,5 @@
 //
-//  DataSourcePipelineDeduplicationTests.swift
+//  RestoreImageMergeTests.swift
 //  BushelCloud
 //
 //  Created by Leo Dion.
@@ -33,72 +33,7 @@ import Testing
 @testable import BushelCloudKit
 @testable import BushelFoundation
 
-// MARK: - Suite 1: RestoreImage Deduplication Tests
-
-@Suite("RestoreImage Deduplication")
-struct RestoreImageDeduplicationTests {
-  let pipeline = DataSourcePipeline()
-
-  @Test("Empty array returns empty")
-  func testDeduplicateEmpty() {
-    let result = pipeline.deduplicateRestoreImages([])
-    #expect(result.isEmpty)
-  }
-
-  @Test("Single record returns unchanged")
-  func testDeduplicateSingle() {
-    let input = [TestFixtures.sonoma14_2_1]
-    let result = pipeline.deduplicateRestoreImages(input)
-
-    #expect(result.count == 1)
-    #expect(result[0].buildNumber == "23C71")
-  }
-
-  @Test("Different builds all preserved")
-  func testDeduplicateDifferentBuilds() {
-    let input = [
-      TestFixtures.sonoma14_2_1,
-      TestFixtures.sequoia15_1,
-      TestFixtures.sonoma14_0,
-    ]
-    let result = pipeline.deduplicateRestoreImages(input)
-
-    #expect(result.count == 3)
-    // Should be sorted by releaseDate descending
-    #expect(result[0].buildNumber == "24B83")  // sequoia15_1 (Nov 2024)
-    #expect(result[1].buildNumber == "23C71")  // sonoma14_2_1 (Dec 2023)
-    #expect(result[2].buildNumber == "23A344")  // sonoma14_0 (Sep 2023)
-  }
-
-  @Test("Duplicate builds merged")
-  func testDeduplicateDuplicateBuilds() {
-    let input = [
-      TestFixtures.sonoma14_2_1,
-      TestFixtures.sonoma14_2_1_mesu,
-      TestFixtures.sonoma14_2_1_appledb,
-    ]
-    let result = pipeline.deduplicateRestoreImages(input)
-
-    // Should have only 1 record after merging
-    #expect(result.count == 1)
-    #expect(result[0].buildNumber == "23C71")
-  }
-
-  @Test("Results sorted by release date descending")
-  func testSortingByReleaseDateDescending() {
-    let input = [
-      TestFixtures.sonoma14_0,  // Oldest: Sep 2023
-      TestFixtures.sonoma14_2_1,  // Middle: Dec 2023
-      TestFixtures.sequoia15_1,  // Newest: Nov 2024
-    ]
-    let result = pipeline.deduplicateRestoreImages(input)
-
-    #expect(result.count == 3)
-    // Verify descending order
-    #expect(result[0].releaseDate > result[1].releaseDate)
-    #expect(result[1].releaseDate > result[2].releaseDate)
-  }
-}
+// swiftlint:disable file_length
 
 // MARK: - Suite 2: RestoreImage Merge Tests
 
@@ -428,226 +363,5 @@ struct RestoreImageMergeTests {
     let merged = pipeline.mergeRestoreImages(first, second)
 
     #expect(merged.notes == "First note; Second note")
-  }
-}
-
-// MARK: - Suite 3: XcodeVersion Reference Resolution Tests
-
-@Suite("XcodeVersion Reference Resolution")
-struct XcodeVersionReferenceResolutionTests {
-  let pipeline = DataSourcePipeline()
-
-  @Test("Resolve exact version match 14.2")
-  func testResolveExactMatch() {
-    let xcode = TestFixtures.xcodeWithRequires_14_2
-    let restoreImages = [TestFixtures.restoreImage_14_2]
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == "RestoreImage-23C64")
-  }
-
-  @Test("Resolve 3-component version 14.2.1")
-  func testResolveThreeComponentVersion() {
-    let xcode = TestFixtures.xcodeWithRequires_14_2_1
-    let restoreImages = [TestFixtures.sonoma14_2_1]
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == "RestoreImage-23C71")
-  }
-
-  @Test("Resolve 2-component to 3-component match")
-  func testResolveTwoToThreeComponent() {
-    let xcode = TestFixtures.xcodeWithRequires_14_2
-    let restoreImages = [TestFixtures.sonoma14_2_1]  // version="14.2.1"
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    // Should match via short version "14.2" to "14.2.1"
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == "RestoreImage-23C71")
-  }
-
-  @Test("No match leaves minimumMacOS nil")
-  func testNoMatchLeavesNil() {
-    let xcode = TestFixtures.xcodeWithRequires_14_2
-    let restoreImages = [TestFixtures.sequoia15_1]  // Different version
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == nil)
-  }
-
-  @Test("No REQUIRES field leaves minimumMacOS nil")
-  func testNoRequiresLeavesNil() {
-    let xcode = TestFixtures.xcodeNoRequires
-    let restoreImages = [TestFixtures.sonoma14_2_1]
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == nil)
-  }
-
-  @Test("Invalid REQUIRES format leaves minimumMacOS nil")
-  func testInvalidRequiresLeavesNil() {
-    let xcode = TestFixtures.xcodeInvalidRequires
-    let restoreImages = [TestFixtures.sonoma14_2_1]
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == nil)
-  }
-
-  @Test("NOTES_URL preserved after resolution")
-  func testNotesURLPreserved() {
-    let xcode = TestFixtures.xcodeWithRequires_14_2
-    let restoreImages = [TestFixtures.restoreImage_14_2]
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].notes == "https://developer.apple.com/notes")
-  }
-
-  @Test("Empty restoreImages array leaves all nil")
-  func testEmptyRestoreImagesArray() {
-    let xcode = TestFixtures.xcodeWithRequires_14_2
-    let restoreImages: [RestoreImageRecord] = []
-
-    let resolved = pipeline.resolveXcodeVersionReferences([xcode], restoreImages: restoreImages)
-
-    #expect(resolved.count == 1)
-    #expect(resolved[0].minimumMacOS == nil)
-  }
-
-  @Test("Multiple Xcodes resolved correctly")
-  func testMultipleXcodeResolution() {
-    let xcodes = [
-      TestFixtures.xcodeWithRequires_14_2,
-      TestFixtures.xcodeWithRequires_14_2_1,
-      TestFixtures.xcodeNoRequires,
-    ]
-    let restoreImages = [
-      TestFixtures.restoreImage_14_2,
-      TestFixtures.sonoma14_2_1,
-    ]
-
-    let resolved = pipeline.resolveXcodeVersionReferences(xcodes, restoreImages: restoreImages)
-
-    #expect(resolved.count == 3)
-    // When both "14.2" and "14.2.1" exist, the short version from "14.2.1"
-    // overwrites "14.2" in the lookup table (last processed wins)
-    // First should resolve to 14.2.1 (due to lookup table collision)
-    #expect(resolved[0].minimumMacOS == "RestoreImage-23C71")
-    // Second should resolve to 14.2.1 (exact match)
-    #expect(resolved[1].minimumMacOS == "RestoreImage-23C71")
-    // Third has no REQUIRES, should remain nil
-    #expect(resolved[2].minimumMacOS == nil)
-  }
-}
-
-// MARK: - Suite 4: XcodeVersion Deduplication Tests
-
-@Suite("XcodeVersion Deduplication")
-struct XcodeVersionDeduplicationTests {
-  let pipeline = DataSourcePipeline()
-
-  @Test("Empty array returns empty")
-  func testDeduplicateEmpty() {
-    let result = pipeline.deduplicateXcodeVersions([])
-    #expect(result.isEmpty)
-  }
-
-  @Test("Single record returns unchanged")
-  func testDeduplicateSingle() {
-    let input = [TestFixtures.xcode15_1]
-    let result = pipeline.deduplicateXcodeVersions(input)
-
-    #expect(result.count == 1)
-    #expect(result[0].buildNumber == "15C65")
-  }
-
-  @Test("Duplicate builds keep first occurrence")
-  func testDuplicateBuildsKeepFirst() {
-    let input = [
-      TestFixtures.xcode15_1,
-      TestFixtures.xcode15_1_duplicate,
-    ]
-    let result = pipeline.deduplicateXcodeVersions(input)
-
-    #expect(result.count == 1)
-    // Should keep first occurrence
-    #expect(result[0].buildNumber == "15C65")
-    #expect(result[0].notes == "Release notes: https://developer.apple.com/xcode/release-notes/")
-  }
-
-  @Test("Results sorted by release date descending")
-  func testSortingByReleaseDateDescending() {
-    let input = [
-      TestFixtures.xcode15_1,  // Dec 2023
-      TestFixtures.xcode16_0,  // Sep 2024
-    ]
-    let result = pipeline.deduplicateXcodeVersions(input)
-
-    #expect(result.count == 2)
-    // Verify descending order (newer first)
-    #expect(result[0].buildNumber == "16A242d")  // xcode16_0
-    #expect(result[1].buildNumber == "15C65")  // xcode15_1
-  }
-}
-
-// MARK: - Suite 5: SwiftVersion Deduplication Tests
-
-@Suite("SwiftVersion Deduplication")
-struct SwiftVersionDeduplicationTests {
-  let pipeline = DataSourcePipeline()
-
-  @Test("Empty array returns empty")
-  func testDeduplicateEmpty() {
-    let result = pipeline.deduplicateSwiftVersions([])
-    #expect(result.isEmpty)
-  }
-
-  @Test("Single record returns unchanged")
-  func testDeduplicateSingle() {
-    let input = [TestFixtures.swift5_9_2]
-    let result = pipeline.deduplicateSwiftVersions(input)
-
-    #expect(result.count == 1)
-    #expect(result[0].version == "5.9.2")
-  }
-
-  @Test("Duplicate versions keep first occurrence")
-  func testDuplicateVersionsKeepFirst() {
-    let input = [
-      TestFixtures.swift5_9_2,
-      TestFixtures.swift5_9_2_duplicate,
-    ]
-    let result = pipeline.deduplicateSwiftVersions(input)
-
-    #expect(result.count == 1)
-    // Should keep first occurrence
-    #expect(result[0].version == "5.9.2")
-    #expect(result[0].notes == "Stable Swift release bundled with Xcode 15.1")
-  }
-
-  @Test("Results sorted by release date descending")
-  func testSortingByReleaseDateDescending() {
-    let input = [
-      TestFixtures.swift5_9_2,  // Dec 2023
-      TestFixtures.swift6_1,  // Nov 2024
-    ]
-    let result = pipeline.deduplicateSwiftVersions(input)
-
-    #expect(result.count == 2)
-    // Verify descending order (newer first)
-    #expect(result[0].version == "6.1")  // swift6_1
-    #expect(result[1].version == "5.9.2")  // swift5_9_2
   }
 }
