+++ Tests/BushelCloudKitTests/DataSources/VirtualBuddyFetcherTests.swift
@@ -37,9 +37,11 @@ import Testing
   import FoundationNetworking
 #endif
 
+// swiftlint:disable file_length type_body_length
+
 /// All VirtualBuddy tests wrapped in a serialized suite to prevent mock handler conflicts
 @Suite("VirtualBuddyFetcher Tests", .serialized)
-struct VirtualBuddyFetcherAllTests {
+struct VirtualBuddyFetcherTests {
   // MARK: - Initialization Tests
 
   @Suite("Initialization")
