+++ Tests/BushelCloudKitTests/Mocks/MockAppleDBFetcher.swift
@@ -1,0 +1,52 @@
+//
+//  MockAppleDBFetcher.swift
+//  BushelCloud
+//
+//  Created by Leo Dion.
+//  Copyright © 2025 BrightDigit.
+//
+//  Permission is hereby granted, free of charge, to any person
+//  obtaining a copy of this software and associated documentation
+//  files (the "Software"), to deal in the Software without
+//  restriction, including without limitation the rights to use,
+//  copy, modify, merge, publish, distribute, sublicense, and/or
+//  sell copies of the Software, and to permit persons to whom the
+//  Software is furnished to do so, subject to the following
+//  conditions:
+//
+//  The above copyright notice and this permission notice shall be
+//  included in all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+//  OTHER DEALINGS IN THE SOFTWARE.
+//
+
+import Foundation
+
+@testable import BushelFoundation
+
+/// Mock fetcher for AppleDB data source
+struct MockAppleDBFetcher: DataSourceFetcher, Sendable {
+  typealias Record = [RestoreImageRecord]
+
+  let recordsToReturn: [RestoreImageRecord]
+  let errorToThrow: (any Error)?
+
+  init(recordsToReturn: [RestoreImageRecord] = [], errorToThrow: (any Error)? = nil) {
+    self.recordsToReturn = recordsToReturn
+    self.errorToThrow = errorToThrow
+  }
+
+  func fetch() async throws -> [RestoreImageRecord] {
+    if let error = errorToThrow {
+      throw error
+    }
+    return recordsToReturn
+  }
+}
