+++ Tests/BushelCloudKitTests/Utilities/TestFixtures.swift
@@ -33,6 +33,7 @@ public import Foundation
 @testable public import BushelFoundation
 
 /// Centralized test data fixtures for all record types
+// swiftlint:disable identifier_name file_length type_body_length
 public enum TestFixtures: Sendable {
   // MARK: - RestoreImage Fixtures
 
@@ -494,3 +495,4 @@ public enum TestFixtures: Sendable {
     }
     """
 }
+// swiftlint:enable identifier_name file_length type_body_length
